// src/rbac/enums/permissions.ts
export enum permissions {
	
	PRODUCT_CREATE = 'PRODUCT_CREATE',
	PRODUCT_READ = 'PRODUCT_READ',
	PRODUCT_UPDATE = 'PRODUCT_UPDATE',
	PRODUCT_DELETE = 'PRODUCT_DELETE',
	PRODUCT_BULK_OPERATIONS = 'PRODUCT_BULK_OPERATIONS',
  

	INVENTORY_READ = 'INVENTORY_READ',
	INVENTORY_UPDATE = 'INVENTORY_UPDATE',
	INVENTORY_BULK_UPDATE = 'INVENTORY_BULK_UPDATE',

	CATEGORY_CREATE = 'CATEGORY_CREATE',
	CATEGORY_READ = 'CATEGORY_READ',
	CATEGORY_UPDATE = 'CATEGORY_UPDATE',
	CATEGORY_DELETE = 'CATEGORY_DELETE',
  
	
	PRODUCT_SEARCH = 'PRODUCT_SEARCH',
	ADVANCED_SEARCH = 'ADVANCED_SEARCH',
  

	IMAGE_UPLOAD = 'IMAGE_UPLOAD',
	IMAGE_DELETE = 'IMAGE_DELETE',
  }
  